--- a/llamafile/iqk_mul_mat.inc
+++ b/llamafile/iqk_mul_mat.inc
@@ -2054,28 +2054,28 @@ struct DequantizerIQ2XXS final : public BaseDequantizer<block_iq2_xxs> {
         prepare2(bits.b2.val + 2, idx, keven_signs, sidx[3]);
     }
 
-    inline void new_block(int i) {
-        d = 0.125f * GGML_FP16_TO_FP32(x[i].d);
-    }
-    inline int32_t process_block(int i, int j, const Q8<1, block_q8_K>& q8) {
-        uint32_t aux32[8];
-        std::memcpy(aux32, x[i].qs + 16*j, 32);
-        auto q1 = q8.load_quants_64(0, i, 2*j+0);
-        auto q2 = q8.load_quants_64(0, i, 2*j+1);
-        int8x16_t * b1 = (int8x16_t *)bits.b1.val;
-        int8x16_t * b2 = (int8x16_t *)bits.b2.val;
-        prepare4(bits.b1.val, aux32+0);
-        prepare4(bits.b2.val, aux32+4);
-        auto p1 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b1[0], q1.val[0]), b1[1], q1.val[1]);
-        auto p2 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b1[2], q1.val[2]), b1[3], q1.val[3]);
-        auto p3 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b2[0], q2.val[0]), b2[1], q2.val[1]);
-        auto p4 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b2[2], q2.val[2]), b2[3], q2.val[3]);
-        int32_t sumi = vaddvq_s32(p1) * (2*(aux32[1] >> 28) + 1)
-                     + vaddvq_s32(p2) * (2*(aux32[3] >> 28) + 1)
-                     + vaddvq_s32(p3) * (2*(aux32[5] >> 28) + 1)
-                     + vaddvq_s32(p4) * (2*(aux32[7] >> 28) + 1);
-        return sumi;
-    }
+    //inline void new_block(int i) {
+    //    d = 0.125f * GGML_FP16_TO_FP32(x[i].d);
+    //}
+    //inline int32_t process_block(int i, int j, const Q8<1, block_q8_K>& q8) {
+    //    uint32_t aux32[8];
+    //    std::memcpy(aux32, x[i].qs + 16*j, 32);
+    //    auto q1 = q8.load_quants_64(0, i, 2*j+0);
+    //    auto q2 = q8.load_quants_64(0, i, 2*j+1);
+    //    int8x16_t * b1 = (int8x16_t *)bits.b1.val;
+    //    int8x16_t * b2 = (int8x16_t *)bits.b2.val;
+    //    prepare4(bits.b1.val, aux32+0);
+    //    prepare4(bits.b2.val, aux32+4);
+    //    auto p1 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b1[0], q1.val[0]), b1[1], q1.val[1]);
+    //    auto p2 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b1[2], q1.val[2]), b1[3], q1.val[3]);
+    //    auto p3 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b2[0], q2.val[0]), b2[1], q2.val[1]);
+    //    auto p4 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b2[2], q2.val[2]), b2[3], q2.val[3]);
+    //    int32_t sumi = vaddvq_s32(p1) * (2*(aux32[1] >> 28) + 1)
+    //                 + vaddvq_s32(p2) * (2*(aux32[3] >> 28) + 1)
+    //                 + vaddvq_s32(p3) * (2*(aux32[5] >> 28) + 1)
+    //                 + vaddvq_s32(p4) * (2*(aux32[7] >> 28) + 1);
+    //    return sumi;
+    //}
 
     uint32x4x4_t data;
     SimpleBits bits;
@@ -2249,37 +2249,37 @@ struct DequantizerIQ3XXS final : public BaseDequantizer<block_iq3_xxs> {
         make2(q3, signs[2], bits.b2.val + 0); q3 += 8;
         make2(q3, signs[3], bits.b2.val + 2);
     }
-    inline void prepare_v2(int i, int j, const uint32_t * signs) {
-        const auto * q3 = x[i].qs + 32*j;
-        make2_unsigned(q3, bits.b1.val + 0); q3 += 8;
-        make2_unsigned(q3, bits.b1.val + 2); q3 += 8;
-        apply_signs_2(bits.b1.val+0, keven_signs, signs[0]);
-        apply_signs_2(bits.b1.val+2, keven_signs, signs[1]);
-        make2_unsigned(q3, bits.b2.val + 0); q3 += 8;
-        make2_unsigned(q3, bits.b2.val + 2);
-        apply_signs_2(bits.b2.val+0, keven_signs, signs[2]);
-        apply_signs_2(bits.b2.val+2, keven_signs, signs[3]);
-    }
+    //inline void prepare_v2(int i, int j, const uint32_t * signs) {
+    //    const auto * q3 = x[i].qs + 32*j;
+    //    make2_unsigned(q3, bits.b1.val + 0); q3 += 8;
+    //    make2_unsigned(q3, bits.b1.val + 2); q3 += 8;
+    //    apply_signs_2(bits.b1.val+0, keven_signs, signs[0]);
+    //    apply_signs_2(bits.b1.val+2, keven_signs, signs[1]);
+    //    make2_unsigned(q3, bits.b2.val + 0); q3 += 8;
+    //    make2_unsigned(q3, bits.b2.val + 2);
+    //    apply_signs_2(bits.b2.val+0, keven_signs, signs[2]);
+    //    apply_signs_2(bits.b2.val+2, keven_signs, signs[3]);
+    //}
 
-    inline void new_block(int i) {
-        d = 0.25f * GGML_FP16_TO_FP32(x[i].d);
-    }
+    //inline void new_block(int i) {
+    //    d = 0.25f * GGML_FP16_TO_FP32(x[i].d);
+    //}
 
-    inline int32_t process_block(int i, int j, const Q8<1, block_q8_K>& q8) {
-        gas.val[0] = vld1q_u32((const uint32_t *)(x[i].qs + QK_K/4 + 16*j));
-        prepare_v2(i, j, (const uint32_t *)gas.val);
-        gas.val[0] = vorrq_u32(vshlq_n_u32(vshrq_n_u32(gas.val[0], 28), 1), vdupq_n_u32(1));
-        int8x16_t * b1 = (int8x16_t *)bits.b1.val;
-        int8x16_t * b2 = (int8x16_t *)bits.b2.val;
-        auto q1 = q8.load_quants_64(0, i, 2*j+0);
-        auto q2 = q8.load_quants_64(0, i, 2*j+1);
-        auto p1 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b1[0], q1.val[0]), b1[1], q1.val[1]);
-        auto p2 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b1[2], q1.val[2]), b1[3], q1.val[3]);
-        auto p3 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b2[0], q2.val[0]), b2[1], q2.val[1]);
-        auto p4 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b2[2], q2.val[2]), b2[3], q2.val[3]);
-        const int32_t * s32 = (const int32_t *)gas.val;
-        return vaddvq_s32(p1) * s32[0] + vaddvq_s32(p2) * s32[1] + vaddvq_s32(p3) * s32[2] + vaddvq_s32(p4) * s32[3];
-    }
+    //inline int32_t process_block(int i, int j, const Q8<1, block_q8_K>& q8) {
+    //    gas.val[0] = vld1q_u32((const uint32_t *)(x[i].qs + QK_K/4 + 16*j));
+    //    prepare_v2(i, j, (const uint32_t *)gas.val);
+    //    gas.val[0] = vorrq_u32(vshlq_n_u32(vshrq_n_u32(gas.val[0], 28), 1), vdupq_n_u32(1));
+    //    int8x16_t * b1 = (int8x16_t *)bits.b1.val;
+    //    int8x16_t * b2 = (int8x16_t *)bits.b2.val;
+    //    auto q1 = q8.load_quants_64(0, i, 2*j+0);
+    //    auto q2 = q8.load_quants_64(0, i, 2*j+1);
+    //    auto p1 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b1[0], q1.val[0]), b1[1], q1.val[1]);
+    //    auto p2 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b1[2], q1.val[2]), b1[3], q1.val[3]);
+    //    auto p3 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b2[0], q2.val[0]), b2[1], q2.val[1]);
+    //    auto p4 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b2[2], q2.val[2]), b2[3], q2.val[3]);
+    //    const int32_t * s32 = (const int32_t *)gas.val;
+    //    return vaddvq_s32(p1) * s32[0] + vaddvq_s32(p2) * s32[1] + vaddvq_s32(p3) * s32[2] + vaddvq_s32(p4) * s32[3];
+    //}
 
     SimpleBits bits;
     uint32x4x2_t gas;
@@ -2340,31 +2340,31 @@ struct DequantizerIQ3S final : public BaseDequantizer<block_iq3_s> {
         make4(sh, x[i].signs + 16*j + 8, qs+16, qh+2, hshift, bits.b2.val);
     }
 
-    inline void new_block(int i) {
-        d = GGML_FP16_TO_FP32(x[i].d);
-    }
-    inline int32_t process_block(int i, int j, const Q8<1, block_q8_K>& q8) {
-        prepare(i, j);
-        const uint16_t * aux16 = (const uint16_t *)x[i].scales + j;
-        uint32_t scales32 = (((aux16[0] | (aux16[0] << 12)) & 0x0f0f0f0f) << 1) | 0x01010101;
-        int8x16_t * b1 = (int8x16_t *)bits.b1.val;
-        int8x16_t * b2 = (int8x16_t *)bits.b2.val;
-        auto q1 = q8.load_quants_64(0, i, 2*j+0);
-        auto q2 = q8.load_quants_64(0, i, 2*j+1);
-        //auto p1 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b1[0], q1.val[0]), b1[1], q1.val[1]);
-        //auto p2 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b1[2], q1.val[2]), b1[3], q1.val[3]);
-        //auto p3 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b2[0], q2.val[0]), b2[1], q2.val[1]);
-        //auto p4 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b2[2], q2.val[2]), b2[3], q2.val[3]);
-        //const int8_t * s8 = (const int8_t *)&scales32;
-        //return vaddvq_s32(p1) * s8[0] + vaddvq_s32(p2) * s8[2] + vaddvq_s32(p3) * s8[1] + vaddvq_s32(p4) * s8[3];
-        auto zero = vdupq_n_s32(0);
-        auto p1 = vaddvq_s32(ggml_vdotq_s32(ggml_vdotq_s32(zero, b1[0], q1.val[0]), b1[1], q1.val[1]));
-        auto p2 = vaddvq_s32(ggml_vdotq_s32(ggml_vdotq_s32(zero, b1[2], q1.val[2]), b1[3], q1.val[3]));
-        auto p3 = vaddvq_s32(ggml_vdotq_s32(ggml_vdotq_s32(zero, b2[0], q2.val[0]), b2[1], q2.val[1]));
-        auto p4 = vaddvq_s32(ggml_vdotq_s32(ggml_vdotq_s32(zero, b2[2], q2.val[2]), b2[3], q2.val[3]));
-        const int8_t * s8 = (const int8_t *)&scales32;
-        return p1 * s8[0] + p2 * s8[2] + p3 * s8[1] + p4 * s8[3];
-    }
+    //inline void new_block(int i) {
+    //    d = GGML_FP16_TO_FP32(x[i].d);
+    //}
+    //inline int32_t process_block(int i, int j, const Q8<1, block_q8_K>& q8) {
+    //    prepare(i, j);
+    //    const uint16_t * aux16 = (const uint16_t *)x[i].scales + j;
+    //    uint32_t scales32 = (((aux16[0] | (aux16[0] << 12)) & 0x0f0f0f0f) << 1) | 0x01010101;
+    //    int8x16_t * b1 = (int8x16_t *)bits.b1.val;
+    //    int8x16_t * b2 = (int8x16_t *)bits.b2.val;
+    //    auto q1 = q8.load_quants_64(0, i, 2*j+0);
+    //    auto q2 = q8.load_quants_64(0, i, 2*j+1);
+    //    //auto p1 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b1[0], q1.val[0]), b1[1], q1.val[1]);
+    //    //auto p2 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b1[2], q1.val[2]), b1[3], q1.val[3]);
+    //    //auto p3 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b2[0], q2.val[0]), b2[1], q2.val[1]);
+    //    //auto p4 = ggml_vdotq_s32(ggml_vdotq_s32(vdupq_n_s32(0), b2[2], q2.val[2]), b2[3], q2.val[3]);
+    //    //const int8_t * s8 = (const int8_t *)&scales32;
+    //    //return vaddvq_s32(p1) * s8[0] + vaddvq_s32(p2) * s8[2] + vaddvq_s32(p3) * s8[1] + vaddvq_s32(p4) * s8[3];
+    //    auto zero = vdupq_n_s32(0);
+    //    auto p1 = vaddvq_s32(ggml_vdotq_s32(ggml_vdotq_s32(zero, b1[0], q1.val[0]), b1[1], q1.val[1]));
+    //    auto p2 = vaddvq_s32(ggml_vdotq_s32(ggml_vdotq_s32(zero, b1[2], q1.val[2]), b1[3], q1.val[3]));
+    //    auto p3 = vaddvq_s32(ggml_vdotq_s32(ggml_vdotq_s32(zero, b2[0], q2.val[0]), b2[1], q2.val[1]));
+    //    auto p4 = vaddvq_s32(ggml_vdotq_s32(ggml_vdotq_s32(zero, b2[2], q2.val[2]), b2[3], q2.val[3]));
+    //    const int8_t * s8 = (const int8_t *)&scales32;
+    //    return p1 * s8[0] + p2 * s8[2] + p3 * s8[1] + p4 * s8[3];
+    //}
 
     SimpleBits bits;
     SignHelper1 sh;
@@ -2433,29 +2433,29 @@ IQK_NOINLINE void mul_mat_qX_K_q8_K_T(int n, const void * vx, size_t bx, const D
     }
 }
 
-template <typename Dequantizer>
-IQK_NOINLINE void mul_mat_qX_K_q8_K_1(int n, const void * vx, size_t bx, const DataInfo& info, int nrc_x) {
-    assert(n % QK_K == 0);
-    const int nb = n / QK_K;
-
-    Q8<1, block_q8_K> q8(info);
-
-    Dequantizer deq(vx, bx, 1);
-
-    for (int ix = 0; ix < nrc_x; ++ix) {
-
-        deq.new_row(ix);
-
-        float32_t acc = 0;
-        for (int i = 0; i < nb; ++i) {
-            deq.new_block(i);
-            auto sumi = deq.process_block(i, 0, q8);
-            sumi += deq.process_block(i, 1, q8);
-            acc += deq.d*q8.scale(0, i)*sumi;
-        }
-        info.store(ix, 0, acc);
-    }
-}
+//template <typename Dequantizer>
+//IQK_NOINLINE void mul_mat_qX_K_q8_K_1(int n, const void * vx, size_t bx, const DataInfo& info, int nrc_x) {
+//    assert(n % QK_K == 0);
+//    const int nb = n / QK_K;
+//
+//    Q8<1, block_q8_K> q8(info);
+//
+//    Dequantizer deq(vx, bx, 1);
+//
+//    for (int ix = 0; ix < nrc_x; ++ix) {
+//
+//        deq.new_row(ix);
+//
+//        float32_t acc = 0;
+//        for (int i = 0; i < nb; ++i) {
+//            deq.new_block(i);
+//            auto sumi = deq.process_block(i, 0, q8);
+//            sumi += deq.process_block(i, 1, q8);
+//            acc += deq.d*q8.scale(0, i)*sumi;
+//        }
+//        info.store(ix, 0, acc);
+//    }
+//}
 
 // =========================================== Legacy quants
 
@@ -2909,18 +2909,18 @@ template <typename Dequantizer> void MulMat::set_functions(MulMat& m) {
         m.funcs[6] = mul_mat_qX_1_q8_1<Dequantizer, 7>;
         m.funcs[7] = mul_mat_qX_1_q8_1<Dequantizer, 8>;
     }
-    else if constexpr (std::is_same_v<Dequantizer, DequantizerIQ2XXS> || std::is_same_v<Dequantizer, DequantizerIQ3S> ||
-                       std::is_same_v<Dequantizer, DequantizerIQ3XXS>) {// || std::is_same_v<Dequantizer, DequantizerIQ2S>) {
-        m.funcs[0] = mul_mat_qX_K_q8_K_1<Dequantizer>;
-        //m.funcs[0] = mul_mat_qX_K_q8_K_T<1, DequantizerIQ2XXS>;
-        m.funcs[1] = mul_mat_qX_K_q8_K_T<2, Dequantizer>;
-        m.funcs[2] = mul_mat_qX_K_q8_K_T<3, Dequantizer>;
-        m.funcs[3] = mul_mat_qX_K_q8_K_T<4, Dequantizer>;
-        m.funcs[4] = mul_mat_qX_K_q8_K_T<5, Dequantizer>;
-        m.funcs[5] = mul_mat_qX_K_q8_K_T<6, Dequantizer>;
-        m.funcs[6] = mul_mat_qX_K_q8_K_T<7, Dequantizer>;
-        m.funcs[7] = mul_mat_qX_K_q8_K_T<8, Dequantizer>;
-    }
+    //else if constexpr (std::is_same_v<Dequantizer, DequantizerIQ2XXS> || std::is_same_v<Dequantizer, DequantizerIQ3S> ||
+    //                   std::is_same_v<Dequantizer, DequantizerIQ3XXS>) {// || std::is_same_v<Dequantizer, DequantizerIQ2S>) {
+    //    m.funcs[0] = mul_mat_qX_K_q8_K_1<Dequantizer>;
+    //    //m.funcs[0] = mul_mat_qX_K_q8_K_T<1, DequantizerIQ2XXS>;
+    //    m.funcs[1] = mul_mat_qX_K_q8_K_T<2, Dequantizer>;
+    //    m.funcs[2] = mul_mat_qX_K_q8_K_T<3, Dequantizer>;
+    //    m.funcs[3] = mul_mat_qX_K_q8_K_T<4, Dequantizer>;
+    //    m.funcs[4] = mul_mat_qX_K_q8_K_T<5, Dequantizer>;
+    //    m.funcs[5] = mul_mat_qX_K_q8_K_T<6, Dequantizer>;
+    //    m.funcs[6] = mul_mat_qX_K_q8_K_T<7, Dequantizer>;
+    //    m.funcs[7] = mul_mat_qX_K_q8_K_T<8, Dequantizer>;
+    //}
     else {
         m.funcs[0] = mul_mat_qX_K_q8_K_T<1, Dequantizer>;
         m.funcs[1] = mul_mat_qX_K_q8_K_T<2, Dequantizer>;